services/grid-normalizer/requirements.txt
aiokafka==0.10.0
psycopg2-binary==2.9.9
pydantic==2.7.1
pydantic-settings==2.2.1
structlog==24.1.0
pytest==8.2.0
pytest-asyncio==0.23.6
